Renamed and Deleted files
Renamed stacked.py to run_cough_detection_gui.py (last version of ui)
Deleted previous uneeded files
