+++ cough_simplerColors_not_stacked.ui
