--- last.py
@@ -1,279 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'last_screen.ui'
-#
-# Created by: PyQt5 UI code generator 5.9.2
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_last_screen(QtWidgets.QWidget):
-    def setupUi(self, MainWindow):
-        MainWindow.setObjectName("MainWindow")
-        MainWindow.resize(532, 603)
-        MainWindow.setFixedSize(self.size())        
-        self.centralwidget = QtWidgets.QWidget(MainWindow)
-        self.centralwidget.setObjectName("centralwidget")
-        self.frame = QtWidgets.QFrame(self.centralwidget)
-        self.frame.setGeometry(QtCore.QRect(0, 0, 531, 600))
-        self.frame.setStyleSheet("background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(221, 231, 199, 255), stop:1 rgba(238, 248, 214, 255));\n"
-"border-radius: 10px\n"
-"")
-        self.frame.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame.setObjectName("frame")
-        self.frame_2 = QtWidgets.QFrame(self.frame)
-        self.frame_2.setGeometry(QtCore.QRect(20, 80, 490, 361))
-        self.frame_2.setStyleSheet("background-color: rgb(227, 227, 227);border-radius:10px;")
-        self.frame_2.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_2.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_2.setObjectName("frame_2")
-        self.wav_folder = QtWidgets.QTextBrowser(self.frame_2)
-        self.wav_folder.setGeometry(QtCore.QRect(230, 10, 251, 31))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.wav_folder.setFont(font)
-        self.wav_folder.setStyleSheet("background-color:rgb(255, 255, 255)")
-        self.wav_folder.setObjectName("wav_folder")
-        self.label = QtWidgets.QLabel(self.frame_2)
-        self.label.setGeometry(QtCore.QRect(10, 10, 221, 21))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(False)
-        font.setUnderline(True)
-        font.setWeight(50)
-        self.label.setFont(font)
-        self.label.setStyleSheet("")
-        self.label.setObjectName("label")
-        self.label_3 = QtWidgets.QLabel(self.frame_2)
-        self.label_3.setGeometry(QtCore.QRect(10, 60, 151, 16))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setUnderline(True)
-        self.label_3.setFont(font)
-        self.label_3.setObjectName("label_3")
-        self.results = QtWidgets.QPlainTextEdit(self.frame_2)
-        self.results.setGeometry(QtCore.QRect(10, 90, 471, 251))
-        self.results.setStyleSheet("background-color: rgb(255, 255, 255);")
-        self.results.setObjectName("results")
-        self.filelist = QtWidgets.QComboBox(self.frame_2)
-        self.filelist.setGeometry(QtCore.QRect(230, 60, 251, 22))
-        self.filelist.setStyleSheet("background-color: rgb(255, 255, 255);")
-        self.filelist.setObjectName("filelist")
-        self.frame_3 = QtWidgets.QFrame(self.frame)
-        self.frame_3.setGeometry(QtCore.QRect(0, 0, 531, 31))
-        self.frame_3.setStyleSheet("background-color: rgba(1, 32, 15, 100)")
-        self.frame_3.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_3.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_3.setObjectName("frame_3")
-        self.btn_min = QtWidgets.QPushButton(self.frame_3)
-        self.btn_min.setGeometry(QtCore.QRect(470, 5, 20, 20))
-        font = QtGui.QFont()
-        font.setFamily("Calibri")
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.btn_min.setFont(font)
-        self.btn_min.setStyleSheet("QPushButton{ \n"
-"border:none; \n"
-"border-radius: 7px; \n"
-"background-color:rgb(240, 135, 0);\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: rgba(240, 135, 0, 150);\n"
-"}")
-        self.btn_min.setObjectName("btn_min")
-        self.btn_min.clicked.connect(self.minimize)
-        self.btn_exit = QtWidgets.QPushButton(self.frame_3)
-        self.btn_exit.setGeometry(QtCore.QRect(500, 5, 20, 20))
-        self.btn_exit.clicked.connect(self.on_closing)        
-        font = QtGui.QFont()
-        font.setFamily("Corbel")
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.btn_exit.setFont(font)
-        self.btn_exit.setStyleSheet("QPushButton{ \n"
-"border:none; \n"
-"border-radius: 7px; \n"
-"background-color: rgb(235, 0, 0);\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: rgb(235,0,0, 150);\n"
-"}")
-        self.btn_exit.setObjectName("btn_exit")
-        self.title = QtWidgets.QLabel(self.frame_3)
-        self.title.setGeometry(QtCore.QRect(11, -3, 331, 41))
-        font = QtGui.QFont()
-        font.setFamily("Caladea")
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setItalic(False)
-        font.setWeight(75)
-        self.title.setFont(font)
-        self.title.setAutoFillBackground(False)
-        self.title.setStyleSheet("background-color: rgba(255, 255, 255, 0);")
-        self.title.setObjectName("title")
-        self.cite_label = QtWidgets.QPlainTextEdit(self.frame)
-        self.cite_label.setGeometry(QtCore.QRect(20, 500, 510, 100))
-        font = QtGui.QFont()
-        font.setFamily("Calibri")
-        font.setPointSize(10)
-        font.setItalic(True)
-        self.cite_label.setFont(font)
-        self.cite_label.setMouseTracking(False)
-        self.cite_label.setReadOnly(True)
-        self.cite_label.setObjectName("cite_label")
-        self.copy_citation = QtWidgets.QPushButton(self.frame)
-        self.copy_citation.setGeometry(QtCore.QRect(390, 570, 121, 21))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.copy_citation.setFont(font)
-        self.copy_citation.setStyleSheet("QPushButton{ \n"
-"border:0.5px solid grey; \n"
-"border-radius: 7px; \n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
-"}")
-        self.copy_citation.setObjectName("copy_citation")
-        self.copy_citation.clicked.connect(self.copy_citation_)        
-        self.btn_xlsx = QtWidgets.QPushButton(self.frame)
-        self.btn_xlsx.setGeometry(QtCore.QRect(300, 460, 151, 31))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.btn_xlsx.setFont(font)
-        self.btn_xlsx.setStyleSheet("QPushButton{ \n"
-"border:0.5px solid grey; \n"
-"border-radius: 7px; \n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
-"}\n"
-"\n"
-"QPushButton:disabled{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 10), stop:1 rgba(255, 255, 255, 10));\n"
-"}\n"
-"\n"
-"")
-        self.btn_xlsx.setObjectName("btn_xlsx")
-        self.btn_xlsx.cicked.connect(export_xlsx)
-        self.btn_txt = QtWidgets.QPushButton(self.frame)
-        self.btn_txt.setGeometry(QtCore.QRect(60, 460, 151, 31))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.btn_txt.setFont(font)
-        self.btn_txt.setStyleSheet("QPushButton{ \n"
-"border:0.5px solid grey; \n"
-"border-radius: 7px; \n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
-"}\n"
-"\n"
-"QPushButton:disabled{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 10), stop:1 rgba(255, 255, 255, 10));\n"
-"}\n"
-"\n"
-"")
-        self.btn_txt.setObjectName("btn_txt")
-        self.btn_txt.cicked.connect(export_txt)
-        self.title_2 = QtWidgets.QLabel(self.frame)
-        self.title_2.setGeometry(QtCore.QRect(30, 40, 81, 41))
-        font = QtGui.QFont()
-        font.setFamily("Caladea")
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setItalic(False)
-        font.setWeight(75)
-        self.title_2.setFont(font)
-        self.title_2.setAutoFillBackground(False)
-        self.title_2.setStyleSheet("background-color: rgba(255, 255, 255, 0);")
-        self.title_2.setObjectName("title_2")
-        MainWindow.setCentralWidget(self.centralwidget)
-
-        self.retranslateUi(MainWindow)
-        QtCore.QMetaObject.connectSlotsByName(MainWindow)
-
-    def retranslateUi(self, MainWindow):
-        _translate = QtCore.QCoreApplication.translate
-        MainWindow.setWindowTitle(_translate("MainWindow", "MainWindow"))
-        self.wav_folder.setHtml(_translate("MainWindow", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.0//EN\" \"http://www.w3.org/TR/REC-html40/strict.dtd\">\n"
-"<html><head><meta name=\"qrichtext\" content=\"1\" /><style type=\"text/css\">\n"
-"p, li { white-space: pre-wrap; }\n"
-"</style></head><body style=\" font-family:\'MS Shell Dlg 2\'; font-size:9pt; font-weight:400; font-style:normal;\">\n"
-"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:7.8pt;\">[Please select a directory]</span></p></body></html>"))
-        self.label.setText(_translate("MainWindow", "Directory with wav files scanned:"))
-        self.label_3.setText(_translate("MainWindow", "Results per file:"))
-        self.btn_min.setText(_translate("MainWindow", "_"))
-        self.btn_exit.setText(_translate("MainWindow", "X"))
-        self.title.setText(_translate("MainWindow", "Cough Detection Tool (v 1.0)"))
-        self.cite_label.setPlainText(_translate("MainWindow", "Did you find this tool useful? Please cite as:\n"
-"Nikonas Simou; Nikolaos Stefanakis; Panagiotis Zervas, \"A Universal System for Cough Detection in Domestic Acoustic Environments\" in EUSIPCO 2020."))
-        self.copy_citation.setText(_translate("MainWindow", "Copy citation"))
-        self.btn_xlsx.setText(_translate("MainWindow", "Save as .xlsx"))
-        self.btn_txt.setText(_translate("MainWindow", "Save as .txt"))
-        self.title_2.setText(_translate("MainWindow", "Results"))
-
-    def minimize(self):
-        self.showMinimized()
-
-    def select_dir(self):
-        import glob
-        folder = str(QtWidgets.QFileDialog.getExistingDirectory(self, "Select Directory"))
-        if folder:
-            self.wav_folder.setText(folder)
-            files = glob.glob(f"{folder}/**/*.wav", recursive=True)
-            if len(files)>0:
-                files = "\n".join(files).replace("\\", "/").replace(folder, "./")
-                self.btn_run.setEnabled(True)
-                self.btn_run.setText("Run")
-            else:
-                files = "No .wav files found in selected directory."
-                self.btn_run.setText("Run\n(Select a valid directory)")
-                self.btn_run.setEnabled(False)
-            self.wavs.setText(files)
-            self.btn_dir.setText("Change\nDirectory")
-        #window.label_2.setText(file)
-
-
-    def on_closing(self):
-        import gc
-
-        reply = QtWidgets.QMessageBox.question(self, 'Message',
-            "Are you sure to quit?", QtWidgets.QMessageBox.Yes | 
-            QtWidgets.QMessageBox.No, QtWidgets.QMessageBox.No)
-
-        if reply == QtWidgets.QMessageBox.Yes:
-            print('Thank you for using our tool!')
-            QtCore.QCoreApplication.instance().quit()
-            gc.collect()
-            exit()
-
-
-    def copy_citation_(self):
-        clipboard = QtWidgets.QApplication.clipboard()
-        clipboard.setText(citation)
-        # self.parent.update() # the text will stay there after the window is closed
-        # self.copied_lbl_txt.set("(Copied!)")
-        
-    def export_xlsx(self):
-        pass
-    def export_txt(self):
-        pass
-
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    MainWindow = QtWidgets.QMainWindow()
-    ui = Ui_MainWindow()
-    ui.setupUi(MainWindow)
-    MainWindow.show()
-    sys.exit(app.exec_())
-
