--- last_screen.py
@@ -1,210 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'last_screen.ui'
-#
-# Created by: PyQt5 UI code generator 5.9.2
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-class Ui_LastScreen(QtWidgets.QWidget):
-    def setupUi(self, MainWindow):
-        MainWindow.setObjectName("MainWindow")
-        MainWindow.resize(530, 610)
-        MainWindow.setFixedSize(self.size())
-        self.centralwidget = QtWidgets.QWidget(MainWindow)
-        self.centralwidget.setObjectName("centralwidget")
-        self.centralwidget.setWindowFlags(QtCore.Qt.FramelessWindowHint)
-        self.frame = QtWidgets.QFrame(self.centralwidget)
-        self.frame.setGeometry(QtCore.QRect(0, 0, 530, 610))
-        self.frame.setStyleSheet("background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(221, 231, 199, 255), stop:1 rgba(238, 248, 214, 255));\n"
-"border-radius: 10px\n"
-"")
-        self.frame.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame.setObjectName("frame")
-        self.frame_2 = QtWidgets.QFrame(self.frame)
-        self.frame_2.setGeometry(QtCore.QRect(20, 50, 490, 391))
-        self.frame_2.setStyleSheet("background-color: rgb(227, 227, 227);border-radius:10px;")
-        self.frame_2.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_2.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_2.setObjectName("frame_2")
-        self.wav_folder = QtWidgets.QTextBrowser(self.frame_2)
-        self.wav_folder.setGeometry(QtCore.QRect(230, 10, 251, 31))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.wav_folder.setFont(font)
-        self.wav_folder.setStyleSheet("background-color:rgb(255, 255, 255)")
-        self.wav_folder.setObjectName("wav_folder")
-        self.label = QtWidgets.QLabel(self.frame_2)
-        self.label.setGeometry(QtCore.QRect(10, 10, 221, 21))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(False)
-        font.setUnderline(True)
-        font.setWeight(50)
-        self.label.setFont(font)
-        self.label.setStyleSheet("")
-        self.label.setObjectName("label")
-        self.label_3 = QtWidgets.QLabel(self.frame_2)
-        self.label_3.setGeometry(QtCore.QRect(10, 50, 151, 16))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setUnderline(True)
-        self.label_3.setFont(font)
-        self.label_3.setObjectName("label_3")
-        self.results = QtWidgets.QPlainTextEdit(self.frame_2)
-        self.results.setGeometry(QtCore.QRect(10, 70, 471, 301))
-        self.results.setStyleSheet("background-color: rgb(255, 255, 255);")
-        self.results.setObjectName("results")
-        self.frame_3 = QtWidgets.QFrame(self.frame)
-        self.frame_3.setGeometry(QtCore.QRect(0, 0, 531, 31))
-        self.frame_3.setStyleSheet("background-color: rgba(1, 32, 15, 100)")
-        self.frame_3.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_3.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_3.setObjectName("frame_3")
-        self.btn_min = QtWidgets.QPushButton(self.frame_3)
-        self.btn_min.setGeometry(QtCore.QRect(470, 5, 20, 20))
-        font = QtGui.QFont()
-        font.setFamily("Calibri")
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.btn_min.setFont(font)
-        self.btn_min.setStyleSheet("QPushButton{ \n"
-"border:none; \n"
-"border-radius: 7px; \n"
-"background-color:rgb(240, 135, 0);\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: rgba(240, 135, 0, 150);\n"
-"}")
-        self.btn_min.setObjectName("btn_min")
-        self.btn_exit = QtWidgets.QPushButton(self.frame_3)
-        self.btn_exit.setGeometry(QtCore.QRect(500, 5, 20, 20))
-        font = QtGui.QFont()
-        font.setFamily("Corbel")
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.btn_exit.setFont(font)
-        self.btn_exit.setStyleSheet("QPushButton{ \n"
-"border:none; \n"
-"border-radius: 7px; \n"
-"background-color: rgb(235, 0, 0);\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: rgb(235,0,0, 150);\n"
-"}")
-        self.btn_exit.setObjectName("btn_exit")
-        self.title = QtWidgets.QLabel(self.frame_3)
-        self.title.setGeometry(QtCore.QRect(11, -3, 331, 41))
-        font = QtGui.QFont()
-        font.setFamily("Caladea")
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setItalic(False)
-        font.setWeight(75)
-        self.title.setFont(font)
-        self.title.setAutoFillBackground(False)
-        self.title.setStyleSheet("background-color: rgba(255, 255, 255, 0);")
-        self.title.setObjectName("title")
-        self.cite_label = QtWidgets.QPlainTextEdit(self.frame)
-        self.cite_label.setGeometry(QtCore.QRect(20, 500, 510, 100))
-        font = QtGui.QFont()
-        font.setFamily("Calibri")
-        font.setPointSize(10)
-        font.setItalic(True)
-        self.cite_label.setFont(font)
-        self.cite_label.setMouseTracking(False)
-        self.cite_label.setReadOnly(True)
-        self.cite_label.setObjectName("cite_label")
-        self.copy_citation_2 = QtWidgets.QPushButton(self.frame)
-        self.copy_citation_2.setGeometry(QtCore.QRect(390, 570, 121, 21))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.copy_citation_2.setFont(font)
-        self.copy_citation_2.setStyleSheet("QPushButton{ \n"
-"border:0.5px solid grey; \n"
-"border-radius: 7px; \n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
-"}")
-        self.copy_citation_2.setObjectName("copy_citation_2")
-        self.btn_xlsx = QtWidgets.QPushButton(self.frame)
-        self.btn_xlsx.setGeometry(QtCore.QRect(300, 460, 151, 31))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.btn_xlsx.setFont(font)
-        self.btn_xlsx.setStyleSheet("QPushButton{ \n"
-"border:0.5px solid grey; \n"
-"border-radius: 7px; \n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
-"}\n"
-"\n"
-"QPushButton:disabled{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 10), stop:1 rgba(255, 255, 255, 10));\n"
-"}\n"
-"\n"
-"")
-        self.btn_xlsx.setObjectName("btn_xlsx")
-        self.btn_txt = QtWidgets.QPushButton(self.frame)
-        self.btn_txt.setGeometry(QtCore.QRect(60, 460, 151, 31))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        self.btn_txt.setFont(font)
-        self.btn_txt.setStyleSheet("QPushButton{ \n"
-"border:0.5px solid grey; \n"
-"border-radius: 7px; \n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(60, 137, 109, 255), stop:1 rgba(60, 137, 109, 100));\n"
-"}\n"
-"QPushButton:hover{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 255), stop:1 rgba(255, 255, 255, 255));\n"
-"}\n"
-"\n"
-"QPushButton:disabled{\n"
-"background-color: qlineargradient(spread:pad, x1:0, y1:0, x2:1, y2:0, stop:0 rgba(172, 194, 170, 10), stop:1 rgba(255, 255, 255, 10));\n"
-"}\n"
-"\n"
-"")
-        self.btn_txt.setObjectName("btn_txt")
-        MainWindow.setCentralWidget(self.centralwidget)
-
-        self.retranslateUi(MainWindow)
-        QtCore.QMetaObject.connectSlotsByName(MainWindow)
-
-    def retranslateUi(self, MainWindow):
-        _translate = QtCore.QCoreApplication.translate
-        MainWindow.setWindowTitle(_translate("MainWindow", "MainWindow"))
-        self.wav_folder.setHtml(_translate("MainWindow", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.0//EN\" \"http://www.w3.org/TR/REC-html40/strict.dtd\">\n"
-"<html><head><meta name=\"qrichtext\" content=\"1\" /><style type=\"text/css\">\n"
-"p, li { white-space: pre-wrap; }\n"
-"</style></head><body style=\" font-family:\'MS Shell Dlg 2\'; font-size:9pt; font-weight:400; font-style:normal;\">\n"
-"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:7.8pt;\">[Please select a directory]</span></p></body></html>"))
-        self.label.setText(_translate("MainWindow", "Directory with wav files scanned:"))
-        self.label_3.setText(_translate("MainWindow", "Results per file:"))
-        self.btn_min.setText(_translate("MainWindow", "_"))
-        self.btn_exit.setText(_translate("MainWindow", "X"))
-        self.title.setText(_translate("MainWindow", "Cough Detection Tool (v 1.0)"))
-        self.cite_label.setPlainText(_translate("MainWindow", "Did you find this tool useful? Please cite as:\n"
-"Nikonas Simou; Nikolaos Stefanakis; Panagiotis Zervas, \"A Universal System for Cough Detection in Domestic Acoustic Environments\" in EUSIPCO 2020."))
-        self.copy_citation_2.setText(_translate("MainWindow", "Copy citation"))
-        self.btn_xlsx.setText(_translate("MainWindow", "Save as .xlsx"))
-        self.btn_txt.setText(_translate("MainWindow", "Save as .txt"))
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    MainWindow = QtWidgets.QMainWindow()
-    ui = Ui_MainWindow()
-    ui.setupUi(MainWindow)
-    MainWindow.show()
-    sys.exit(app.exec_())
-
