+++ run_cough_detection_gui.ui
@@ -24,7 +24,7 @@
      </rect>
     </property>
     <property name="currentIndex">
-     <number>1</number>
+     <number>0</number>
     </property>
     <widget class="QWidget" name="SettingsPage">
      <widget class="QFrame" name="frame">
@@ -302,7 +302,7 @@ border-radius: 7px;
 background-color: rgb(235, 0, 0);
 }
 QPushButton:hover{
-background-color: rgb(235,0,0, 150);
+background-color: rgba(235,0,0, 150);
 }</string>
         </property>
         <property name="text">
@@ -676,7 +676,7 @@ border-radius: 7px;
 background-color: rgb(235, 0, 0);
 }
 QPushButton:hover{
-background-color: rgb(235,0,0, 150);
+background-color: rgba(235,0,0, 150);
 }</string>
         </property>
         <property name="text">
